editor: Remove TheGraph global from clipboard and menu code
Instead use regular CommonJS require(),
making cross-module dependencies clearer.
